+++ R/learner_glmnet_surv_cv_glmnet.R
@@ -1,6 +1,7 @@
 #' @title Cross-Validated GLM with Elastic Net Regularization Survival Learner
-#' @author be-marc
 #' @name mlr_learners_surv.cv_glmnet
+#' @author be-marc
+#' @author bblodfon
 #'
 #' @description
 #' Generalized linear models with elastic net regularization.
@@ -8,6 +9,8 @@
 #'
 #' @section Initial parameter values:
 #' - `family` is set to `"cox"` and cannot be changed.
+#' - `cox.ties` is initialized to `"breslow"` to keep the tie-handling behavior of earlier glmnet versions,
+#' and to silence the glmnet v5.0 warning about the upcoming default change to `"efron"`.
 #'
 #' @section Prediction types:
 #' This learner returns three prediction types:
@@ -17,7 +20,7 @@
 #' 2. `crank`: same as `lp`.
 #' 3. `distr`: a survival matrix in two dimensions, where observations are
 #' represented in rows and time points in columns.
-#' Calculated using `glmnet::survfit.cv.glmnet()`.
+#' Calculated using `glmnet:::survfit.cv.glmnet()`.
 #' Parameters `stype` and `ctype` relate to how `lp` predictions are transformed
 #' into survival predictions and are described in `survival::survfit.coxph()`.
 #' By default the Breslow estimator is used for computing the baseline hazard.
@@ -33,60 +36,73 @@
 #' @export
 #' @template seealso_learner
 #' @template example
-LearnerSurvCVGlmnet = R6Class("LearnerSurvCVGlmnet",
+LearnerSurvCVGlmnet = R6Class(
+  "LearnerSurvCVGlmnet",
   inherit = mlr3proba::LearnerSurv,
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      # fmt: skip
       ps = ps(
-        alignment            = p_fct(c("lambda", "fraction"), default = "lambda", tags = "train"),
-        alpha                = p_dbl(0, 1, default = 1, tags = "train"),
-        big                  = p_dbl(default = 9.9e35, tags = "train"),
-        devmax               = p_dbl(0, 1, default = 0.999, tags = "train"),
-        dfmax                = p_int(0L, tags = "train"),
-        eps                  = p_dbl(0, 1, default = 1.0e-6, tags = "train"),
-        epsnr                = p_dbl(0, 1, default = 1.0e-8, tags = "train"),
-        exclude              = p_uty(tags = "train"),
-        exmx                 = p_dbl(default = 250.0, tags = "train"),
-        fdev                 = p_dbl(0, 1, default = 1.0e-5, tags = "train"),
-        foldid               = p_uty(default = NULL, tags = "train"),
-        gamma                = p_uty(tags = "train"),
-        grouped              = p_lgl(default = TRUE, tags = "train"),
-        intercept            = p_lgl(default = TRUE, tags = "train"),
-        keep                 = p_lgl(default = FALSE, tags = "train"),
-        lambda               = p_uty(tags = "train"),
-        lambda.min.ratio     = p_dbl(0, 1, tags = "train"),
-        lower.limits         = p_uty(default = -Inf, tags = "train"),
-        maxit                = p_int(1L, default = 100000L, tags = "train"),
-        mnlam                = p_int(1L, default = 5L, tags = "train"),
-        mxit                 = p_int(1L, default = 100L, tags = "train"),
-        mxitnr               = p_int(1L, default = 25L, tags = "train"),
-        nfolds               = p_int(3L, default = 10L, tags = "train"),
-        nlambda              = p_int(1L, default = 100L, tags = "train"),
-        use_pred_offset      = p_lgl(default = TRUE, tags = "predict"),
-        parallel             = p_lgl(default = FALSE, tags = "train"),
-        penalty.factor       = p_uty(tags = "train"),
-        pmax                 = p_int(0L, tags = "train"),
-        pmin                 = p_dbl(0, 1, default = 1.0e-9, tags = "train"),
-        prec                 = p_dbl(default = 1e-10, tags = "train"),
-        predict.gamma        = p_dbl(default = "gamma.1se", special_vals = list("gamma.1se", "gamma.min"), tags = "predict"),
-        relax                = p_lgl(default = FALSE, tags = "train"),
-        s                    = p_dbl(0, special_vals = list("lambda.1se", "lambda.min"), default = "lambda.1se", tags = "predict"), #nolint
-        standardize          = p_lgl(default = TRUE, tags = "train"),
-        standardize.response = p_lgl(default = FALSE, tags = "train"),
-        thresh               = p_dbl(0, default = 1e-07, tags = "train"),
-        trace.it             = p_int(0, 1, default = 0, tags = "train"),
-        type.gaussian        = p_fct(c("covariance", "naive"), tags = "train"),
-        type.logistic        = p_fct(c("Newton", "modified.Newton"), default = "Newton", tags = "train"),
-        type.measure         = p_fct(c("deviance", "C"), default = "deviance", tags = "train"),
-        type.multinomial     = p_fct(c("ungrouped", "grouped"), default = "ungrouped", tags = "train"),
-        upper.limits         = p_uty(default = Inf, tags = "train"),
-        stype                = p_int(default = 2L, lower = 1L, upper = 2L, tags = "predict"), # default: Breslow
-        ctype                = p_int(lower = 1L, upper = 2L, tags = "predict") # how to handle ties
+        # glmnet::cv.glmnet() parameters
+        lambda           = p_uty(default = NULL, tags = "train"),
+        type.measure     = p_fct(c("deviance", "C"), default = "deviance", tags = "train"),
+        nfolds           = p_int(3L, default = 10L, tags = "train"),
+        foldid           = p_uty(default = NULL, tags = "train"),
+        alignment        = p_fct(c("lambda", "fraction"), default = "lambda", tags = "train"),
+        grouped          = p_lgl(default = TRUE, tags = "train"),
+        keep             = p_lgl(default = FALSE, tags = "train"),
+        parallel         = p_lgl(default = FALSE, tags = "train"),
+        gamma            = p_uty(default = c(0, 0.25, 0.5, 0.75, 1), tags = "train", depends = quote(relax == TRUE)),
+        relax            = p_lgl(default = FALSE, tags = "train"),
+        trace.it         = p_int(0, 1, default = 0, tags = "train"), # alias: itrace
+        # glmnet::glmnet() parameters
+        alpha            = p_dbl(0, 1, default = 1, tags = "train"),
+        nlambda          = p_int(1L, default = 100L, tags = "train"),
+        lambda.min.ratio = p_dbl(0, 1, tags = "train"),
+        standardize      = p_lgl(default = TRUE, tags = "train"),
+        intercept        = p_lgl(default = TRUE, tags = "train"),
+        exclude          = p_uty(default = NULL, tags = "train"),
+        penalty.factor   = p_uty(tags = "train"),
+        lower.limits     = p_uty(default = -Inf, tags = "train"),
+        upper.limits     = p_uty(default = Inf, tags = "train"),
+        cox.ties         = p_fct(c("breslow", "efron"), default = "breslow", tags = "train"),
+        # glmnet::relax.glmnet() parameters
+        maxp             = p_int(1L, tags = "train"),
+        path             = p_lgl(default = FALSE, tags = "train"),
+        # glmnet::glmnet.control() parameters
+        fdev             = p_dbl(0, 1, default = 1.0e-5, tags = "train"),
+        devmax           = p_dbl(0, 1, default = 0.999, tags = "train"),
+        eps              = p_dbl(0, 1, default = 1.0e-6, tags = "train"),
+        big              = p_dbl(default = 9.9e+35, tags = "train"),
+        mnlam            = p_int(default = 5L, tags = "train"),
+        pmin             = p_dbl(0, 1, default = 1.0e-9, tags = "train"),
+        exmx             = p_dbl(default = 250, tags = "train"),
+        prec             = p_dbl(default = 1e-10, tags = "train"),
+        mxit             = p_int(1L, default = 100L, tags = "train"),
+        epsnr            = p_dbl(0, 1, default = 1.0e-6, tags = "train"),
+        mxitnr           = p_int(1L, default = 25L, tags = "train"),
+        thresh           = p_dbl(0, default = 1e-07, tags = "train"),
+        maxit            = p_int(1L, default = 100000L, tags = "train"),
+        dfmax            = p_int(0L, default = NULL, special_vals = list(NULL), tags = "train"),
+        pmax             = p_int(0L, default = NULL, special_vals = list(NULL), tags = "train"),
+        # glmnet::predict.cv.glmnet() and glmnet::predict.cv.relaxed() parameters
+        s                = p_dbl(0, special_vals = list("lambda.1se", "lambda.min"), default = "lambda.1se", tags = "predict"),
+        predict.gamma    = p_dbl(0, 1, default = "gamma.1se", special_vals = list("gamma.1se", "gamma.min"), tags = "predict"), # renamed from 'gamma' to avoid duplication
+        # glmnet::predict.coxnet() parameters
+        exact            = p_lgl(default = FALSE, tags = "predict"),
+        # glmnet::survfit.coxnet() parameters => survfit.coxph() parameters for distr prediction
+        stype            = p_int(default = 2L, lower = 1L, upper = 2L, tags = "predict"), # default: Breslow
+        ctype            = p_int(lower = 1L, upper = 2L, tags = "predict"), # how to handle ties
+        # for using the offset during prediction
+        use_pred_offset  = p_lgl(init = TRUE, tags = "predict")
       )
 
-      ps$set_values(use_pred_offset = TRUE)
+      # TODO: Remove `cox.ties` initialization once glmnet >= 5.1 defaults to
+      # cox.ties = "efron" without warnings.
+      # Setting now explicitly to avoid warnings.
+      ps$set_values(cox.ties = "breslow")
 
       super$initialize(
         id = "surv.cv_glmnet",
@@ -113,6 +129,14 @@ LearnerSurvCVGlmnet = R6Class("LearnerSurvCVGlmnet",
     }
   ),
 
+  active = list(
+    #' @field native_model (`cv.glmnet`)\cr
+    #' The fitted model.
+    native_model = function() {
+      self$model$model
+    }
+  ),
+
   private = list(
     .train = function(task) {
       data = as.matrix(task$data(cols = task$feature_names))
@@ -136,22 +160,28 @@ LearnerSurvCVGlmnet = R6Class("LearnerSurvCVGlmnet",
       newdata = as.matrix(ordered_features(task, self))
       pv = self$param_set$get_values(tags = "predict")
       pv = rename(pv, "predict.gamma", "gamma")
+      # one predict method requires numeric gamma always
+      if (inherits(self$native_model, "cv.relaxed") && is.character(pv$gamma)) {
+        pv$gamma = self$native_model$relaxed[[pv$gamma]] # gamma.1se or gamma.min
+      }
       pv = glmnet_set_offset(task, "predict", pv)
 
       # get survival matrix
-      fit = invoke(survival::survfit,
-                   formula = self$model$model,
-                   x = self$model$x,
-                   y = self$model$y,
-                   weights = self$model$weights,
-                   offset = self$model$offset,
-                   newx = newdata,
-                   se.fit = FALSE,
-                   .args = pv)
+      fit = invoke(
+        survival::survfit,
+        formula = self$native_model,
+        x = self$model$x,
+        y = self$model$y,
+        weights = self$model$weights,
+        offset = self$model$offset,
+        newx = newdata,
+        se.fit = FALSE,
+        .args = pv
+      )
 
       # get linear predictor
       lp = as.numeric(
-        invoke(predict, self$model$model, newx = newdata, type = "link", .args = pv)
+        invoke(predict, self$native_model, newx = newdata, type = "link", .args = pv)
       )
 
       mlr3proba::surv_return(times = fit$time, surv = t(fit$surv), lp = lp)