+++ tests/testthat/test_paramtest_glmnet_surv_glmnet.R
@@ -1,22 +1,25 @@
 skip_if_not_installed("mlr3proba")
+skip_if_not_installed("survival")
 skip_if_not_installed("glmnet")
 
 test_that("surv.glmnet", {
   learner = lrn("surv.glmnet")
-  fun = list(glmnet::glmnet, glmnet::cv.glmnet, glmnet::glmnet.control)
+  fun = list(glmnet::glmnet, glmnet::relax.glmnet, glmnet::glmnet.control)
   exclude = c(
     "x", # handled by mlr3
     "y", # handled by mlr3
+    "family", # only "cox" is applicable for this learner
     "weights", # handled by mlr3
-    "nfolds", # not used by learner
-    "foldid", # not used by learner
-    "family", # only coxnet available
-    "type.gaussian", # not used by learner
-    "type.measure", # only used by cv.glmnet
-    "standardize.response", # for 'mgaussian' only
+    "offset", # handled by mlr3
+    "fit", # fit object is passed on to relax.glmnet()
+    "check.args", # default TRUE is good for mlr3, no need to expose
+    "type.gaussian", # not applicable for cox family
+    "type.logistic", # not applicable for cox family
+    "standardize.response", # not applicable for cox family
+    "type.multinomial", # not applicable for cox family
+    "control", # individual control params are set directly
     "itrace", # supported via param trace.it
-    "factory", # only used in scripts, no effect within mlr3
-    "offset" # handled by mlr3
+    "factory" # only used in scripts, no effect within mlr3
   )
 
   paramtest = run_paramtest(learner, fun, exclude, tag = "train")
@@ -25,16 +28,15 @@ test_that("surv.glmnet", {
 
 test_that("predict surv.glmnet", {
   learner = lrn("surv.glmnet")
-  fun = glmnet::predict.glmnet
+  fun = list(glmnet::predict.glmnet, glmnet::predict.relaxed, survival::survfit)
   exclude = c(
     "object", # handled via mlr3
     "newx", # handled via mlr3
     "type", # handled via mlr3
-    "predict.gamma", # renamed from gamma
-    "offset", # for distr prediction
+    "newoffset", # handled by mlr3
+    "formula", # handled via mlr3
     "stype", # for distr prediction
     "ctype", # for distr prediction
-    "newoffset", # handled by mlr3
     "use_pred_offset" # for using the offset during prediction
   )